Refactor input and vartests
