--- starrotator/data/demo_planet.txt
@@ -1,7 +1,0 @@
-3.153     a/Rs
-0.0         e
-0.0       omega
-86.79         Inclination
--84.8          Obliquity
-0.08228      Rp/Rs
-1.4811235    Period
