+++ starrotator/data/demo_system.txt
@@ -1,0 +1,18 @@
+sma_Rs          3.153    
+e		0.0
+omega           0.0
+inclination     86.79
+obliquity       -84.8
+RpRs            0.08228
+P          	1.4811235
+veq            	114000.0
+stelinc       	90.0
+drr             0.0
+T            	10000.0
+FeH             0.0
+logg            4.0
+u1              0.93
+u2              -0.23
+mus             0
+R               115000
+model      	PHOENIX
